// src/types/sanity.types.ts
export type Author = {
  _id: string;
  _type: "author";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  slug?: Slug;
  image?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  bio?: {
    children?: {
      marks?: string[];
      text?: string;
      _type: "span";
      _key: string;
    }[];
    style?: "normal";
    listItem?: never;
    markDefs?: {
      href?: string;
      _type: "link";
      _key: string;
    }[];
    level?: number;
    _type: "block";
    _key: string;
  }[];
};
// ...
export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};
// ...
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;